- upgrade to version 0.2.1 of mzdb-* - some improvement in IP prediction

+++ ProlineStudio-mzScope/src/main/java/fr/proline/mzscope/ui/AbstractRawFilePanel.java
@@ -444,9 +444,9 @@ public abstract class AbstractRawFilePanel extends JPanel implements IRawFilePlo
     }
 
     @Override
-    public void displayScan(int index) {
+    public void displayScan(long index) {
         if ((currentScan == null) || (index != currentScan.getIndex())) {
-            currentScan = getCurrentRawfile().getScan(index);
+            currentScan = getCurrentRawfile().getScan((int)index);
             if (currentScan != null) {
                 currentScanTime = currentScan.getRetentionTime();
                 currentScanMarker.setValue(currentScanTime / 60.0);
