+++ ProlineStudio-mzScope/src/main/java/fr/proline/mzscope/ui/IRawFilePlot.java
@@ -27,7 +27,8 @@ public interface IRawFilePlot {
    void extractChromatogram(double minMz, double maxMz);
 
    void displayFeature(Feature f);
-   void displayScan(int index);
+   
+   void displayScan(long index);
 
    IRawFile getCurrentRawfile();
